# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KigaliWeShare API — neighbor-to-neighbor carpool matching for Kigali commuters
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/me:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the current user's profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "404": { description: No profile yet (needs onboarding) }
    put:
      operationId: upsertMyProfile
      tags: [profile]
      summary: Create or update the current user's profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertProfileBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
  /profile/me/vehicle:
    put:
      operationId: upsertMyVehicle
      tags: [profile]
      summary: Create or update the driver's vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertVehicleBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
  /profile/{userId}:
    get:
      operationId: getUserProfile
      tags: [profile]
      summary: Public-ish profile for another user (rating, completed trips, vehicle)
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicProfile" }
        "404": { description: Not found }
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReportBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
  /admin/reports:
    get:
      operationId: listAdminReports
      tags: [admin]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, reviewing, resolved, dismissed, all]
            default: open
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminReport" }
  /admin/reports/{id}:
    put:
      operationId: updateAdminReport
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateReportBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminReport" }
...
components:
  schemas:
...
    Gender:
      type: string
      enum: [female, male, other, prefer_not_to_say]
    Role:
      type: string
      enum: [rider, driver, both]
    AccountStatus:
      type: string
      enum: [pending, verified, suspended]
...
    FuelType:
      type: string
      enum: [petrol, diesel]
    Language:
      type: string
      enum: [en, fr, rw]
    ReportStatus:
      type: string
      enum: [open, reviewing, resolved, dismissed]
...
    Vehicle:
      type: object
      required: [id, make, model, color, plate, seats, fuelType]
      properties:
        id: { type: string }
        make: { type: string }
        model: { type: string }
        color: { type: string }
        plate: { type: string }
        seats: { type: integer, minimum: 1, maximum: 4 }
        photoUrl: { type: string, nullable: true }
        nationalId: { type: string, nullable: true }
        fuelType: { $ref: "#/components/schemas/FuelType" }
        consumptionLPer100Km: { type: number, nullable: true }
    UpsertVehicleBody:
      type: object
      required: [make, model, color, plate, seats]
      properties:
        make: { type: string }
        model: { type: string }
        color: { type: string }
        plate: { type: string }
        seats: { type: integer, minimum: 1, maximum: 4 }
        photoUrl: { type: string, nullable: true }
        nationalId: { type: string, nullable: true }
        fuelType: { $ref: "#/components/schemas/FuelType" }
        consumptionLPer100Km: { type: number, nullable: true }
    Profile:
      type: object
      required: [userId, fullName, gender, role, neighborhoodId, status, createdAt]
      properties:
        userId: { type: string }
        fullName: { type: string }
        gender: { $ref: "#/components/schemas/Gender" }
        role: { $ref: "#/components/schemas/Role" }
        neighborhoodId: { type: string }
        homePickupPoint: { type: string, nullable: true }
        employer: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        emergencyContactName: { type: string, nullable: true }
        emergencyContactPhone: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/AccountStatus" }
        avatarUrl: { type: string, nullable: true }
        averageRating: { type: number, nullable: true }
        completedTrips: { type: integer }
        vehicle:
          oneOf:
            - { $ref: "#/components/schemas/Vehicle" }
            - { type: "null" }
        inviteLabel: { type: string, nullable: true }
        isAdmin: { type: boolean }
        idVerified: { type: boolean }
        preferredLanguage: { $ref: "#/components/schemas/Language" }
        blockedAt: { type: string, format: date-time, nullable: true }
        blockedReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    PublicProfile:
      type: object
      required: [userId, fullName, role, completedTrips]
      properties:
        userId: { type: string }
        fullName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        avatarUrl: { type: string, nullable: true }
        averageRating: { type: number, nullable: true }
        completedTrips: { type: integer }
        neighborhoodName: { type: string, nullable: true }
        gender: { $ref: "#/components/schemas/Gender" }
        vehicle:
          oneOf:
            - { $ref: "#/components/schemas/Vehicle" }
            - { type: "null" }
        idVerified: { type: boolean }
    UpsertProfileBody:
      type: object
      required: [fullName, gender, role, neighborhoodId, phone]
      properties:
        fullName: { type: string }
        gender: { $ref: "#/components/schemas/Gender" }
        role: { $ref: "#/components/schemas/Role" }
        neighborhoodId: { type: string }
        homePickupPoint: { type: string, nullable: true }
        employer: { type: string, nullable: true }
        phone: { type: string }
        emergencyContactName: { type: string, nullable: true }
        emergencyContactPhone: { type: string, nullable: true }
        preferredLanguage: { $ref: "#/components/schemas/Language" }
...
    CreateReportBody:
      type: object
      required: [reportedUserId, reason]
      properties:
        reportedUserId: { type: string }
        tripId: { type: string, nullable: true }
        reason: { type: string }
        details: { type: string, nullable: true }
    Report:
      type: object
      required: [id, reportedUserId, reason, status, createdAt]
      properties:
        id: { type: string }
        reportedUserId: { type: string }
        tripId: { type: string, nullable: true }
        reason: { type: string }
        status: { $ref: "#/components/schemas/ReportStatus" }
        createdAt: { type: string, format: date-time }
    AdminReport:
      type: object
      required: [id, reporterId, reportedUserId, reason, status, createdAt]
      properties:
        id: { type: string }
        reporterId: { type: string }
        reporterName: { type: string }
        reportedUserId: { type: string }
        reportedName: { type: string }
        tripId: { type: string, nullable: true }
        reason: { type: string }
        details: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ReportStatus" }
        resolvedAt: { type: string, format: date-time, nullable: true }
        resolvedByUserId: { type: string, nullable: true }
        resolutionNote: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    UpdateReportBody:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/ReportStatus" }
        resolutionNote: { type: string, nullable: true }
        blockUser: { type: boolean }
...